requirements.txt
requests
beautifulsoup4
PyYAML
icalendar